requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==23.1.0
chardet==4.0.0
cycler==0.11.0
discord.py==1.7.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.4.3
multidict==6.0.4
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
seaborn==0.12.2
six==1.16.0
typing_extensions==4.6.3
tzdata==2023.3
yarl==1.9.2